add CRUD System in controller but don	 work add in table

+++ src/main/java/com/maslyak/practicRest/repository/CatRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatRepository extends JpaRepository<Cat, Long>{
+public interface CatRepository extends JpaRepository<Cat, Integer>{
 }
